Database interface completed, but don't use as is
I realised as I finished writing the tests that because of how I've structured the queries, there's a race condition because of having a shared client.  What could happen is that one transaction begins and a bunch of other random calls are using the same client, so they become part of the transaction unintentionally.  The solution that I'm going to go with is implementing a PGPool and grabbing a client per method of the Database class