Fix failing test by returning only volumes in the EC2 region for which sweepVolumes was called.
This is what the actual call to runaws will return. If we did not match this expectation in the test, the handleVolumeData function will be called for each volume for *every* region, instead of just being called once for each volume (i.e. in the region to which the volume belongs). This isn't the prettiest solution, since we need to specify all of the parameters. This led to extract the parameters passed to runaws for 'describeVolumes' into a class field, so it could be accessed by the
test.