Set default return statement for stubs in beforeEach
Many of tests do not rely on the output of a particular stub but still need to specify an output to avoid errors. This results in a lot of duplication of empty object initializations. One way around this is to specify a default return statement for each of the stubs, and then only explicitly set custom return values an individual test.