Create sweepVolumes method to fetch and process EBS volumes data (#31)
* Added sweepVolumes() that requests info about EBS volumes.
- This commit is actually a cleanup of the commits made in this pull request: https://github.com/taskcluster/ec2-manager/pull/30
- The tests are partially implemented.
- The suggestions in that PR will be implemented in a future PR.

* Implemented EBS volume tests
- Also, added await for async functions
- The last test, "should call handleVolumeData once per volume", does not work
- For more info, see https://github.com/taskcluster/ec2-manager/pull/30

* Fix failing test by returning only volumes in the EC2 region for which sweepVolumes was called.

This is what the actual call to runaws will return. If we did not match this expectation in the test, the handleVolumeData function will be called for each volume for *every* region, instead of just being called once for each volume (i.e. in the region to which the volume belongs). This isn't the prettiest solution, since we need to specify all of the parameters. This led to extract the parameters passed to runaws for 'describeVolumes' into a class field, so it could be accessed by the
test.

* Implement minor changes from code review

* Prepend "private" function names sweepVolumes and handleVolumeData with underscore to indicate they should not be called elsewhere
* Replace explicit arguments in the new tests with a Sinon custom matcher. The custom matcher appears to only support a single parameter (additional parameters are undefined), which corresponds to the first parameter of describeVolumesStub. Luckily, the first parameter is the ec2 instance being requested, so this is sufficient (although not very obvious). It's probably worth looking into a better way to handle this.
* Replace `region` with a hard-coded string in the last test to make it more consistent.

* Set default return statement for stubs in beforeEach

Many of tests do not rely on the output of a particular stub but still need to specify an output to avoid errors. This results in a lot of duplication of empty object initializations. One way around this is to specify a default return statement for each of the stubs, and then only explicitly set custom return values an individual test.