Calculate and update total EBS volume counts and GB usage per region, type, and status (#41)
* Added sweepVolumes() that requests info about EBS volumes.
- This commit is actually a cleanup of the commits made in this pull request: https://github.com/taskcluster/ec2-manager/pull/30
- The tests are partially implemented.
- The suggestions in that PR will be implemented in a future PR.

* Implemented EBS volume tests
- Also, added await for async functions
- The last test, "should call handleVolumeData once per volume", does not work
- For more info, see https://github.com/taskcluster/ec2-manager/pull/30

* Fix failing test by returning only volumes in the EC2 region for which sweepVolumes was called.

This is what the actual call to runaws will return. If we did not match this expectation in the test, the handleVolumeData function will be called for each volume for *every* region, instead of just being called once for each volume (i.e. in the region to which the volume belongs). This isn't the prettiest solution, since we need to specify all of the parameters. This led to extract the parameters passed to runaws for 'describeVolumes' into a class field, so it could be accessed by the
test.

* Implement minor changes from code review

* Prepend "private" function names sweepVolumes and handleVolumeData with underscore to indicate they should not be called elsewhere
* Replace explicit arguments in the new tests with a Sinon custom matcher. The custom matcher appears to only support a single parameter (additional parameters are undefined), which corresponds to the first parameter of describeVolumesStub. Luckily, the first parameter is the ec2 instance being requested, so this is sufficient (although not very obvious). It's probably worth looking into a better way to handle this.
* Replace `region` with a hard-coded string in the last test to make it more consistent.

* Set default return statement for stubs in beforeEach

Many of tests do not rely on the output of a particular stub but still need to specify an output to avoid errors. This results in a lot of duplication of empty object initializations. One way around this is to specify a default return statement for each of the stubs, and then only explicitly set custom return values an individual test.

* I implemented tabulating the data (volume type, size, and state) and calculating the totals for each object.
The test, "should call handleVolumeData exactly once per volume", fails if we don't check thtat the volume is not undefined.
However, this test should have initialized the volumes so it shouldn't be undefined.

* Minor reorganization

* Fixed several logic errors in tabulating the required data and calculating the totals.
Note: The returned totals are an object in the shape: {$type: {count: INT, gb: INT}} which are later placed in outcome info.
The resulting outcome info: {$region: {$volumes: {$vol_type: {count: INT, gb: INT}}}}

* Refactored _handleVolumeData()

* Just moved combined a return statement

* Removed volumes from outcomeInfo

* Paginate data returned by AWS

* Add missing update to hasNextToken value when NextToken is null

* Set MaxResults to max value for describeVolumes endpoint

It appears that specifying no value for MaxResults will return everything, but we are paging for now just to be safe.

* Don't use strict type-checking on the request's NextToken because undefined value should not trigger further polling

* Refactor _handleVolumeData and add references to helper methods for updating statsum series

* Rather than dealing with each volume individually, it makes sense to pass an array containing information about all volumes in the region to _handleVolumeData to make tabulating and submitting the totals easier.
* Helper method names reference "volume" so they can be differentiated by future helper methods which will have similar functionality, when we eventually break up the logic for handling instances and spot requests.

* Add functionality to submit totals to statsum series

In order to update the correct series, it is necessary to know--and be able to rely on--the format of the data. I added many assert statements to ensure that we do not accidentally send incorrect information and muck up the series data.

* Only call _handleVolumeData if the region has any volumes. Comment out assert statements in _updateVolumeTotals for now.

* Clean-up code per suggestions from PR

* Use destructuring assignment for active and unused
* Replace while with do while and remove hasNextToken boolean

* Move code that calculates total and updates statsum series to sweep function, after forEach loop

* Modify the _updateVolumeTotals function to loop over all regions, add 'this' keyword to _handleVolumeData call

* Fix existing tests to match new code format; add empty response from _calculateVolumeTotals

* Fixed lint error

* Changed _calculateVolumeTotals() to return totals in the updated structure (i.e. Separated by type and active/unused)
Modified _submitVolumeTotals() to directly use the total gb and count values to update the global totals (instead of another nested loop)
Commented out this.monitor.count() because this appears to be undefined in _submitVolumeTotals()
Added the line 'const util = require('util');' to be able to print the entire object with console.log(): e.g. console.log('volume ' + util.inspect(tabulatedVolumes[volume], {showHidden: false, depth: null}));

* _calculateVolumeTotals() is organized by type as well now

* Removed use of this object in _submitVolumeTotals()
Removed .only() in tests
Fixed lint issues

* Initialize all regions in totalVolumes  in _calculateVolumeTotals() instead of checking whether the region is in totalVolumes before initializing.

* Added tests for _calculateVolumeTotals()

* Changed foreach() loops to using anonymous function; this object inside _submitVolumeTotals() is now defined

* Added test for volumes of different regions

* I forgot to get rid of the .only

* Use 'measure' rather than 'count' to track metrics, define prefix 'ebs-volumes' to make things look nicer, replace ':' with '.' for consistency

Also, change ids in test so they are unique.

* Add 'type' qualifier to statsum metric name for clarity.

* Add assert statements back to _submitVolumeTotals

Since it isn't possible to programatically check that the statsum series are being updated correctly, use assert statements to verify the input is formatted as expected.

* Store combined active and unused GB usage and volume count unless otherwise specified

* Moved spies to inside sandbox so that sandbox will restore() the spies for us

* Moved spies to inside sandbox so that sandbox will restore() the spies for us

* Removed util module