Bug 1428068: Handle EC2 prices endpoint pagination
The DescribeSpotPriceHistory endpoint returns a field call NextToken
that, if not null, indicates there are additional data to fetch.

On next call, NextToken should be an input parameters to tell the server
we are fetching the remaining data.

Notice it had to move DescribeAvailabilityZones out of the loop because
it doesn't do pagination.