Move node version and change some deployment settings
There's a couple changes we need because of our Heroku environment.  One
of them is to go back to using Node 8.11.4.  While we wish to use Node
10 here, in advance of iid-verify, there's a bug somewhere which breaks
the Postgres library.

The main reason for Node 10 was for a stable N-Api, but it turns out
that Node 8.12 will bump N-Api to stable status per:

https://github.com/nodejs/node/pull/21593